ESP32 LVGL library from v8.3.5 to v8.3.6 (no functional change) (#18342)
* ESP32 LVGL library from v8.3.5 to v8.3.6 (no functional change)

* Add changelog

+++ lib/libesp32_lvgl/lvgl/src/extra/libs/gif/gifdec.h
@@ -29,7 +29,7 @@ typedef struct gd_GIF {
     int32_t anim_start;
     uint16_t width, height;
     uint16_t depth;
-    uint16_t loop_count;
+    int32_t loop_count;
     gd_GCE gce;
     gd_Palette *palette;
     gd_Palette lct, gct;
